Move into model package. Remove member variables and methods to put into it's own class.
